Have client assume other players are frozen until the server says they aren't, fix server asymmetry bug with PLAYER_DIED
git-svn-id: https://legesmotus.svn.sourceforge.net/svnroot/legesmotus/legesmotus/trunk@1121 a5a5d5d7-22d1-44d1-b37a-67d3d99fcfd7

+++ common/Player.hpp
@@ -139,7 +139,7 @@ namespace LM {
 		virtual void set_gun_rotation_degrees(float gun_rotation);
 		void set_gun_rotation_radians(float gun_rotation);
 		virtual void set_is_invisible(bool is_invisible);
-		virtual void set_is_frozen(bool is_frozen, int freeze_time = 0);
+		virtual void set_is_frozen(bool is_frozen, int64_t freeze_time = 0);
 		virtual void set_freeze_time(long freeze_time);
 		virtual void set_is_grabbing_obstacle(bool);
 		virtual void set_current_weapon_id(long id);
